Create VehicleComponent
Now the physics system handles vehicle movement events.
Physics system was cleaned up a little, VehicleDesc class was moved
to its own file.

+++ src/components/CameraComponent.h
@@ -18,7 +18,7 @@ public:
 	void setPerspectiveCamera(float fov, float aspectRatio, float nearPlane, float farPlane);
 	void updateOrthographicCamera(float width, float height, float nearPlane, float farPlane);
 	void updatePerspectiveCamera(float fov, float aspectRatio, float nearPlane, float farPlane);
-	void windowSizeChanged(float width, float height) { camera->windowSizeChanged(width, height); }
+	void windowSizeChanged(int width, int height) { camera->windowSizeChanged(width, height); }
 	CameraPurpose getPurpose() { return purpose; }
 	glm::mat4 getViewMatrix(glm::mat4 cameraModelTransformation);
 	glm::mat4 getProjectionMatrix();
